feat: preserve float64 through the image data path
unwrap keeps float64 sources (numpy dtype / nibabel header storage dtype)
instead of downcasting; standardize and the nyul mapping cast to the
data's own dtype; ravel returns corrected images in the source dtype.
Everything else canonicalizes to float32 as before — integer dtypes
cannot round-trip a normalization by construction. Derived/internal data
(RAVEL's image matrix, membership maps, learned transform parameters)
stays float32 for memory.

New property test: float64 in -> float64 out for all individual methods;
metadata law updated (float64 preserved, else float32 stored).

+++ src/intensity_normalization/methods/nyul.py
@@ -61,7 +61,7 @@ class NyulTransform(FittedTransform):
             self.standard_scale,
             fill_value="extrapolate",
         )
-        return mapping(data).astype(np.float32)
+        return mapping(data).astype(data.dtype)
 
     def _state_dict(self) -> dict[str, np.ndarray]:
         return {
